+++ src/PHPHtmlParser/Enum/StringToken.php
@@ -11,6 +11,7 @@ use MyCLabs\Enum\Enum;
  * @method static StringToken EQUAL()
  * @method static StringToken SLASH()
  * @method static StringToken ATTR()
+ * @method static StringToken CLOSECOMMENT()
  */
 class StringToken extends Enum
 {
@@ -18,4 +19,5 @@ class StringToken extends Enum
     private const EQUAL = ' =/>';
     private const SLASH = " />\r\n\t";
     private const ATTR = ' >';
+    private const CLOSECOMMENT = '-->';
 }